+++ CHANGELOG.md
@@ -23,6 +23,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Fixed duplicated minor number bug (based on orphaned .res file)
 - Fixed reading wrong property for cache's blocksize
 - Fixed incorrect calculation of dm-cache's meta-device size
+- Fixed cases where DrbdVlmData's exists boolean was not properly updated
 
 ## [1.31.0] - 2025-04-08